rename db_cleanup to db_utils
